# Optimize ZSTD_decodeSequence for clang