@Cyan4973 2021-05-21
FYI, early results,
I started a quick series of tests based on this PR,
unfortunately, the outcome is not favorable so far.

The only compiler that seems to register a _tiny_ benefit is `clang-11`,
all other compilers, and versions, get either the same, or slightly worse performance.

This is not final evaluation, I will need more time to properly evaluate.

In any case, it was a good idea to try out.