@danlark1 2021-05-21
Yeah, I also noticed that clang-11 benefits the most. And on some processors there is no seen benefit at all.

Anyway, if you think that's not justified, I am ok, I have several other more global ideas but they need time :)